Fix Dashboard timeline granularity issues
Backend (web/app.py):
- Change hours parameter type from int to float
- Fixes 0.5 hours (30 minutes) being truncated to 0
- Now correctly returns minute-level data for short time ranges

Frontend (Dashboard.vue):
- Add granularity detection (minute/hour/day level)
- Dynamic x-axis labels based on time range:
  - Minutes: show "HH:MM"
  - Hours: show "HH:00"
  - Days: show "MMM DD"
- Recreate chart when switching granularity levels
- Adjust maxTicksLimit based on granularity (6/8/7)
- Add maxRotation for minute labels to prevent overlap

+++ web/app.py
@@ -214,7 +214,7 @@ def api_statistics_summary():
 @app.route("/api/statistics/timeline")
 def api_statistics_timeline():
     """Get timeline statistics with optional time filter."""
-    hours = request.args.get("hours", None, type=int)
+    hours = request.args.get("hours", None, type=float)
     days = request.args.get("days", None, type=int)
     timeline = stats_engine.get_request_volume_timeline(hours=hours, days=days)
     return jsonify(timeline)
